Avoid compiler error with -Werror=format-security
When GCC is run with -Werror=format-security, it aborts the compilation because
it considers printf(ctime(&t)) at risk for string format injection. This should
not be possible anyway when using the unchanged output of ctime()), but we can
avoid the compiler warning by providing a format string and using ctime(&t)
merely as an argument.

Original report and proposed fix:
- https://aur.archlinux.org/packages/vitetris#comment-817850
- https://aur.archlinux.org/packages/vitetris#comment-833456